+++ scaffolding/crud-script.js
@@ -41,8 +41,19 @@ var schema = {
 };
 
 prompt.start();
-
+console.log('############################################################');
+console.log('##                                                        ##');
+console.log('##            #  # ####   #### ###  #  # ###              ##');
+console.log('##            # #    #    #    #  # #  # #   #            ##');
+console.log('##            ##    #  ## #    ###  #  # #   #            ##');
+console.log('##            # #  #      #    # #  #  # #   #            ##');
+console.log('##            #  # ####   #### #  # #### ###              ##');
+console.log('##                                                        ##');
+console.log('##    Seja bem vindo ao gerador de CRUD da Kazale.com     ##');
+console.log('############################################################');
+console.log('');
 console.log('Digite o nome do módulo a ser criado (ex.: pessoa):');
+
 
 prompt.get(schema, function (err, input) {
 	modulo = input.modulo;
@@ -54,55 +65,55 @@ prompt.get(schema, function (err, input) {
 	var pathOrigem = toAbsolutePath('source');
 	var pathDestino = pathOrigem.replace('scaffolding/source', 
 		'src/app/' + modulo);
-	console.log(pathOrigem);
-	console.log(pathDestino);
+	//copia os arquivos para o projeto
 	copydir.sync(pathOrigem, pathDestino);
-
 	//renomeia os arquivos
-	fs.rename(pathDestino + '/usuarios.routes.ts', 
-		pathDestino + '/' + moduloPlural + '.routes.ts');
+	fs.renameSync(pathDestino + '/usuario-routing.module.ts', 
+		pathDestino + '/' + modulo + '-routing.module.ts');
+	fs.renameSync(pathDestino + '/usuario.module.ts', 
+		pathDestino + '/' + modulo + '.module.ts');
 	//cadastrar
-	fs.rename(pathDestino + '/cadastrar/usuario-cadastrar.component.css', 
+	fs.renameSync(pathDestino + '/cadastrar/usuario-cadastrar.component.css', 
 		pathDestino + '/cadastrar/' + modulo + '-cadastrar.component.css');
-	fs.rename(pathDestino + '/cadastrar/usuario-cadastrar.component.html', 
+	fs.renameSync(pathDestino + '/cadastrar/usuario-cadastrar.component.html', 
 		pathDestino + '/cadastrar/' + modulo + '-cadastrar.component.html');
-	fs.rename(pathDestino + '/cadastrar/usuario-cadastrar.component.ts', 
+	fs.renameSync(pathDestino + '/cadastrar/usuario-cadastrar.component.ts', 
 		pathDestino + '/cadastrar/' + modulo + '-cadastrar.component.ts');
-	fs.rename(pathDestino + '/cadastrar/usuario-cadastrar.component.spec.ts', 
+	fs.renameSync(pathDestino + '/cadastrar/usuario-cadastrar.component.spec.ts', 
 		pathDestino + '/cadastrar/' + modulo + '-cadastrar.component.spec.ts');
 	//editar
-	fs.rename(pathDestino + '/editar/usuario-editar.component.css', 
+	fs.renameSync(pathDestino + '/editar/usuario-editar.component.css', 
 		pathDestino + '/editar/' + modulo + '-editar.component.css');
-	fs.rename(pathDestino + '/editar/usuario-editar.component.html', 
+	fs.renameSync(pathDestino + '/editar/usuario-editar.component.html', 
 		pathDestino + '/editar/' + modulo + '-editar.component.html');
-	fs.rename(pathDestino + '/editar/usuario-editar.component.ts', 
+	fs.renameSync(pathDestino + '/editar/usuario-editar.component.ts', 
 		pathDestino + '/editar/' + modulo + '-editar.component.ts');
-	fs.rename(pathDestino + '/editar/usuario-editar.component.spec.ts', 
+	fs.renameSync(pathDestino + '/editar/usuario-editar.component.spec.ts', 
 		pathDestino + '/editar/' + modulo + '-editar.component.spec.ts');
 	//listar
-	fs.rename(pathDestino + '/listar/usuario-listar.component.css', 
+	fs.renameSync(pathDestino + '/listar/usuario-listar.component.css', 
 		pathDestino + '/listar/' + modulo + '-listar.component.css');
-	fs.rename(pathDestino + '/listar/usuario-listar.component.html', 
+	fs.renameSync(pathDestino + '/listar/usuario-listar.component.html', 
 		pathDestino + '/listar/' + modulo + '-listar.component.html');
-	fs.rename(pathDestino + '/listar/usuario-listar.component.ts', 
+	fs.renameSync(pathDestino + '/listar/usuario-listar.component.ts', 
 		pathDestino + '/listar/' + modulo + '-listar.component.ts');
-	fs.rename(pathDestino + '/listar/usuario-listar.component.spec.ts', 
+	fs.renameSync(pathDestino + '/listar/usuario-listar.component.spec.ts', 
 		pathDestino + '/listar/' + modulo + '-listar.component.spec.ts');
 	//visualizar
-	fs.rename(pathDestino + '/visualizar/usuario-visualizar.component.css', 
+	fs.renameSync(pathDestino + '/visualizar/usuario-visualizar.component.css', 
 		pathDestino + '/visualizar/' + modulo + '-visualizar.component.css');
-	fs.rename(pathDestino + '/visualizar/usuario-visualizar.component.html', 
+	fs.renameSync(pathDestino + '/visualizar/usuario-visualizar.component.html', 
 		pathDestino + '/visualizar/' + modulo + '-visualizar.component.html');
-	fs.rename(pathDestino + '/visualizar/usuario-visualizar.component.ts', 
+	fs.renameSync(pathDestino + '/visualizar/usuario-visualizar.component.ts', 
 		pathDestino + '/visualizar/' + modulo + '-visualizar.component.ts');
-	fs.rename(pathDestino + '/visualizar/usuario-visualizar.component.spec.ts', 
+	fs.renameSync(pathDestino + '/visualizar/usuario-visualizar.component.spec.ts', 
 		pathDestino + '/visualizar/' + modulo + '-visualizar.component.spec.ts');
 	//shared
-	fs.rename(pathDestino + '/shared/usuario.model.ts', 
+	fs.renameSync(pathDestino + '/shared/usuario.model.ts', 
 		pathDestino + '/shared/' + modulo + '.model.ts');
-	fs.rename(pathDestino + '/shared/usuario.service.ts', 
+	fs.renameSync(pathDestino + '/shared/usuario.service.ts', 
 		pathDestino + '/shared/' + modulo + '.service.ts');
-	fs.rename(pathDestino + '/shared/usuario.service.spec.ts', 
+	fs.renameSync(pathDestino + '/shared/usuario.service.spec.ts', 
 		pathDestino + '/shared/' + modulo + '.service.spec.ts');
 
 	//replace
@@ -216,14 +227,20 @@ prompt.get(schema, function (err, input) {
 	  console.error('Erro:', error);
 	}
 
-	console.log('**********************************************************');
-	console.log('Adicione as seguintes linhas no arquivo src/app/app.routes.ts');
+	console.log('############################################################');
+	console.log('## CRUD ' + modulo + ' criado com sucesso!!!');
+	console.log('## Execute as operações abaixo para registrar o novo módulo na aplicação:');
+	console.log('#');
+	console.log('## Adicione as seguintes linhas no arquivo src/app/app-routing.module.ts');
+	console.log('#   import { ' + moduloCap + 'Routes } from \'./' + modulo + ';');
+	console.log('#   ...' + moduloCap + 'Routes');
+	console.log('#');
+	console.log('## Importe o módulo no arquivo src/app/app.module.ts');
+	console.log('#   import { ' + moduloCap + 'Module } from \'./' + modulo + ';');
+	console.log('#   ' + moduloCap + 'Module' + ' (em @NgModule({ imports: [] })');
+	console.log('#');
+	console.log('## Acesse o CRUD em: http://localhost:8080/' + moduloPlural);
+	console.log('############################################################');
 	console.log('');
-	console.log('import { ' + moduloCapPlural + 'Routes } from \'./' + modulo + '/' + moduloPlural + '.routes\';');
-	console.log('');
-	console.log('...' + moduloCapPlural + 'Routes');
-	console.log('');
-	console.log('Importe os componentes no arquivo src/app/app.module.ts');
-	console.log('**********************************************************');
 });
 // Fim prompt para obter o nome do módulo