+++ scaffolding/source/cadastrar/usuario-cadastrar.component.ts
@@ -2,7 +2,7 @@
  * Componente de cadastro de #MODULO_PLURAL#.
  *
  * @author Márcio Casale de Souza <contato@kazale.com>
- * @since 0.0.1
+ * @since 0.0.3
  */
 
 /*#import { Component } from '@angular/core';
@@ -45,6 +45,6 @@ export class #MODULO_CAP#CadastrarComponent implements OnInit {
 	/*#cadastrar() {
 		this.#MODULO#.id = new Date().getTime();
 		this.#MODULO#Service.cadastrar(this.#MODULO#);
-		this.router.navigate(['/#MODULO#-listar']);
+		this.router.navigate(['/#MODULO_PLURAL#']);
 	}
 }#*/