+++ scaffolding/source/editar/usuario-editar.component.ts
@@ -2,7 +2,7 @@
  * Componente de edição de #MODULO#.
  *
  * @author Márcio Casale de Souza <contato@kazale.com>
- * @since 0.0.1
+ * @since 0.0.3
  */
 
 /*#import { Component } from '@angular/core';
@@ -48,6 +48,6 @@ export class #MODULO_CAP#EditarComponent implements OnInit {
 	 */
 	/*#atualizar() {
 		this.#MODULO#Service.atualizar(this.id, this.#MODULO#);
-		this.router.navigate(['/#MODULO#-listar']);
+		this.router.navigate(['/#MODULO_PLURAL#']);
 	}
 }#*/