+++ scaffolding/source/index.ts
@@ -3,4 +3,5 @@ export * from './editar';
 export * from './listar';
 export * from './visualizar';
 export * from './shared';
-export * from './#MODULO_PLURAL#.routes';#*/
+export * from './#MODULO#-routing.module';
+export * from './#MODULO#.module';#*/