+++ scaffolding/source/shared/usuario.model.ts
@@ -2,7 +2,7 @@
  * Model de #MODULO_PLURAL#.
  *
  * @author Márcio Casale de Souza <contato@kazale.com>
- * @since 0.0.1
+ * @since 0.0.3
  */
 
 /*#export class #MODULO_CAP# {