+++ scaffolding/source/usuario-routing.module.ts
@@ -1,0 +1,23 @@
+/**
+ * Arquivo de configuração de rotas.
+ *
+ * @author Márcio Casale de Souza <contato@kazale.com>
+ * @since 0.0.3
+ */
+
+/*#import { Routes } from '@angular/router'; 
+
+import { 
+	#MODULO_CAP#ListarComponent,
+	#MODULO_CAP#CadastrarComponent,
+	#MODULO_CAP#EditarComponent,
+	#MODULO_CAP#VisualizarComponent
+} from './';
+
+export const #MODULO_CAP#Routes: Routes = [
+	{ path: '#MODULO_PLURAL#', redirectTo: '#MODULO_PLURAL#/listar' },
+	{ path: '#MODULO_PLURAL#/listar', component: #MODULO_CAP#ListarComponent }, 
+	{ path: '#MODULO_PLURAL#/cadastrar', component: #MODULO_CAP#CadastrarComponent }, 
+	{ path: '#MODULO_PLURAL#/editar/:id', component: #MODULO_CAP#EditarComponent },
+	{ path: '#MODULO_PLURAL#/visualizar/:id', component: #MODULO_CAP#VisualizarComponent }
+];#*/