+++ scaffolding/source/usuario.module.ts
@@ -1,0 +1,32 @@
+/**
+ * Arquivo de configuração do módulo.
+ *
+ * @author Márcio Casale de Souza <contato@kazale.com>
+ * @since 0.0.3
+ */
+
+/*#import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+import { 
+	#MODULO_CAP#CadastrarComponent,
+	#MODULO_CAP#EditarComponent,
+	#MODULO_CAP#ListarComponent,
+	#MODULO_CAP#VisualizarComponent
+} from './';
+
+import { SharedModule } from '../shared';
+
+@NgModule({
+	imports: [ 
+		RouterModule,
+		SharedModule
+	],
+	declarations: [
+		#MODULO_CAP#CadastrarComponent,
+		#MODULO_CAP#EditarComponent,
+		#MODULO_CAP#ListarComponent,
+		#MODULO_CAP#VisualizarComponent
+	]
+})
+export class #MODULO_CAP#Module {}#*/