+++ src/app/demo/index.ts
@@ -1,2 +1,3 @@
 export * from './demo.component';
-export * from './demo.routes';
+export * from './demo-routing.module';
+export * from './demo.module';