+++ src/app/shared/index.ts
@@ -3,4 +3,4 @@ export * from './utils';
 export * from './validators';
 export * from './directives';
 export * from './components';
-
+export * from './shared.module';