+++ src/vendor.ts
@@ -21,5 +21,3 @@ import 'rxjs';
 // Adicione outras bibliotecas aqui, como jQuery, Lodash, ou Bootstrap...
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.min.js';
-
-import 'moment/min/moment.min.js';